AP2/samples/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ap2-samples"
version = "0.1"
description = "Samples demonstrating the Agent Payments Protocol"
dependencies = [
    "a2a-sdk",
    "x402-a2a @ git+https://github.com/google-agentic-commerce/a2a-x402.git#subdirectory=python/x402_a2a",
    "absl-py",
    "flask",
    "flask-cors",
    "google-adk",
    "google-genai",
    "httpx",
    "requests",
    "ap2",
    "cryptography",
    "base58",
    "mcp",
    "starlette",
    "uvicorn[standard]",
    "pyjwt[crypto]",
]

keywords = ["payments", "a2a", "ap2"]
readme = "README.md"
requires-python = ">=3.10"

[project.optional-dependencies]
test = ["pytest", "pytest-asyncio"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
ap2 = { workspace = true }
